backend/requirements.txt
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
pymongo==4.6.1
certifi==2024.2.2
# Google Gemini client
google-generativeai==0.8.3
# Production: gunicorn with uvicorn workers (ASGI)
gunicorn==21.2.0
# FastAPI (primary HTTP API)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
# LangGraph + RAG + agent orchestration
langgraph>=0.2.55
langchain-core>=0.3.28
langchain-google-genai>=2.0.0
crewai>=0.86.0
chromadb>=0.5.0
# RAG quality eval (optional: python scripts/ragas_eval.py)
ragas>=0.1.9
datasets>=2.14.0
